pyproject.toml
[build-system]
requires = ["setuptools>=70"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tidal-hqp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "tidalapi>=0.8",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "httpx>=0.27",
    "pytest-mock>=3.14",
    "coverage[toml]>=7.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tidal_hqp*"]

[tool.pytest.ini_options]
testpaths     = ["tests"]
python_files  = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra -q --tb=short"

[tool.coverage.run]
source  = ["tidal_hqp"]
omit    = ["tests/*", "server.py"]

[tool.coverage.report]
show_missing = true
fail_under   = 80
